Moved new_inspire_api to inspire_api
